requirements.txt
mysql-connector-python
scikit-learn
nltk
networkx
json
random
csv
spacy
flask
openai
transformers
torch
torchvision
mysqlclient
tensorflow
gym
keras
interpreter
matplotlib
numpy
pandas
requests
beautifulsoup4
selenium
pytesseract
pillow
opencv-python
pyttsx3
pyaudio
pydub
pywhatkit
pyautogui
pyjokes
pyttsx3
